Tilføjede brug af extern unzip ved zipfile usuporteret kompresionsformat.
Og ryddede lidt op.
