+++ .gitignore
@@ -1,3 +1,3 @@
-/J2/output/
-/J2/temp/
 /.idea/
+/output/
+/temp/
