+++ data/minfil1.txt
