+++ data/minfil3.txt
