+++ data/minfil6.txt
