+++ data/minfil7.txt
