+++ kompress.py
